Moved SwiftRT helper functions into separate file, added hacky gather operation.
